Before starting the feature implementation, some refactoring needed to be done
